CANTERAS/requirements.txt
Fiona==1.9.3
flatbuffers==23.5.26
folium==0.14.0
fonttools==4.39.2
frozenlist==1.3.3
gast==0.4.0
geographiclib==2.0
geojson==3.0.1
geopandas==0.12.2
geopy==2.3.0
glfw==2.5.3
google-auth==2.21.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.56.0
h5py==3.9.0
idna==3.4
Jinja2==3.1.2
joblib==1.2.0
keras==2.13.1
kiwisolver==1.4.4
libclang==16.0.0
Markdown==3.4.3
MarkupSafe==2.1.2
matplotlib==3.7.1
multidict==6.0.4
multipledispatch==0.6.0
munch==2.5.0
networkx==3.1
npm==0.1.1
numpy==1.24.3
oauthlib==3.2.2
openai==0.27.2
openpyxl==3.1.2
opt-einsum==3.3.0
optional-django==0.1.0
osmnx==1.3.0
packaging==23.1
pandas==1.5.3
pefile==2023.2.7
Pillow==9.4.0
plotly==5.13.1
protobuf==4.23.4
psycopg2==2.9.5
psycopg2-binary==2.9.5
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycodestyle==2.8.0
pyinstaller==5.8.0
pyinstaller-hooks-contrib==2023.0
PyOpenGL==3.1.6
PyOpenGL-accelerate==3.1.6
pyparsing==3.1.1
pyproj==3.5.0
PyQt5==5.15.6
PyQt5-Qt5==5.15.2
PyQt5-sip==12.10.1
pyrr==0.10.3
python-dateutil==2.8.2
pytz==2022.7.1
pywin32-ctypes==0.2.0
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.2.2
scipy==1.10.1
shapely==2.0.1
six==1.16.0
sklearn==0.0.post5
sqlparse==0.4.2
tenacity==8.2.2
tensorboard==2.13.0
tensorboard-data-server==0.7.1
tensorflow==2.13.0
tensorflow-estimator==2.13.0
tensorflow-intel==2.13.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
threadpoolctl==3.1.0
toml==0.10.2
tqdm==4.65.0
typing_extensions==4.5.0
tzdata==2022.1
urllib3==1.26.14
vincenty==0.1.4
Werkzeug==2.3.6
wrapt==1.15.0
yarl==1.8.2
